+++ pkg/storage/tsdb/caching_bucket_test.go
@@ -49,6 +49,76 @@ func Test_ChunkCacheBackendValidation(t *testing.T) {
 			},
 			expectedErr: errUnsupportedChunkCacheBackend,
 		},
+		"valid multi chunk cache type": {
+			cfg: ChunkCacheBackend{
+				Backend: fmt.Sprintf("%s,%s,%s", CacheBackendInMemory, CacheBackendMemcached, CacheBackendRedis),
+				Memcached: MemcachedClientConfig{
+					Addresses: "dns+localhost:11211",
+				},
+				Redis: RedisClientConfig{
+					Addresses: "localhost:6379",
+				},
+				MultiLevel: MultiLevelChunkCacheConfig{
+					MaxAsyncConcurrency: 1,
+					MaxAsyncBufferSize:  1,
+					MaxBackfillItems:    1,
+				},
+			},
+			expectedErr: nil,
+		},
+		"duplicate multi chunk cache type": {
+			cfg: ChunkCacheBackend{
+				Backend: fmt.Sprintf("%s,%s", CacheBackendInMemory, CacheBackendInMemory),
+				MultiLevel: MultiLevelChunkCacheConfig{
+					MaxAsyncConcurrency: 1,
+					MaxAsyncBufferSize:  1,
+					MaxBackfillItems:    1,
+				},
+			},
+			expectedErr: errDuplicatedChunkCacheBackend,
+		},
+		"invalid max async concurrency": {
+			cfg: ChunkCacheBackend{
+				Backend: fmt.Sprintf("%s,%s", CacheBackendInMemory, CacheBackendMemcached),
+				Memcached: MemcachedClientConfig{
+					Addresses: "dns+localhost:11211",
+				},
+				MultiLevel: MultiLevelChunkCacheConfig{
+					MaxAsyncConcurrency: 0,
+					MaxAsyncBufferSize:  1,
+					MaxBackfillItems:    1,
+				},
+			},
+			expectedErr: errInvalidMaxAsyncConcurrency,
+		},
+		"invalid max async buffer size": {
+			cfg: ChunkCacheBackend{
+				Backend: fmt.Sprintf("%s,%s", CacheBackendInMemory, CacheBackendMemcached),
+				Memcached: MemcachedClientConfig{
+					Addresses: "dns+localhost:11211",
+				},
+				MultiLevel: MultiLevelChunkCacheConfig{
+					MaxAsyncConcurrency: 1,
+					MaxAsyncBufferSize:  0,
+					MaxBackfillItems:    1,
+				},
+			},
+			expectedErr: errInvalidMaxAsyncBufferSize,
+		},
+		"invalid max back fill items": {
+			cfg: ChunkCacheBackend{
+				Backend: fmt.Sprintf("%s,%s", CacheBackendInMemory, CacheBackendMemcached),
+				Memcached: MemcachedClientConfig{
+					Addresses: "dns+localhost:11211",
+				},
+				MultiLevel: MultiLevelChunkCacheConfig{
+					MaxAsyncConcurrency: 1,
+					MaxAsyncBufferSize:  1,
+					MaxBackfillItems:    0,
+				},
+			},
+			expectedErr: errInvalidMaxBackfillItems,
+		},
 	}
 
 	for name, tc := range tests {